test: move IdeDiagnostics tests
